ProfileView: layout update
