--- src/components/profileCards/index.js
@@ -1,17 +1,0 @@
-import NameCard from './NameCard';
-import EmailCard from './EmailCard';
-import PasswordCard from './PasswordCard';
-import LocationCard from './LocationCard';
-import PhoneCard from './PhoneCard';
-import HomeTelephoneCard from './HomeTelephoneCard';
-import PrivacyCard from './PrivacyCard';
-
-export {
-    NameCard,
-    EmailCard,
-    PasswordCard,
-    LocationCard,
-    PhoneCard,
-    HomeTelephoneCard,
-    PrivacyCard,
-};
